requirements.txt
annotated-types==0.7.0
anyio==4.6.0
av==12.3.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
coloredlogs==15.0.1
ctranslate2==4.4.0
distro==1.9.0
faster-whisper==1.0.3
filelock==3.16.1
flatbuffers==24.3.25
fsspec==2024.9.0
google-ai-generativelanguage==0.6.10
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.2
googleapis-common-protos==1.65.0
groq==0.11.0
grpcio==1.66.2
grpcio-status==1.66.2
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.25.1
humanfriendly==10.0
idna==3.10
jiter==0.5.0
mpmath==1.3.0
numpy==2.1.1
onnxruntime==1.19.2
openai==1.51.0
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.1.4
pyperclip==1.9.0
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
setuptools==75.1.0
sniffio==1.3.1
SpeechRecognition==3.10.4
sympy==1.13.3
tokenizers==0.20.0
tqdm==4.66.5
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
